GET/POST for login and register
when creating an account, will store the user .json into the 'scratch' folder if it exists, creates it if not. Will bring the user to the main page upon successful login/registration

Note: Checks to see if password contains \", or is empty/consists of only whitespace.

Requires: node-localStorage, body-parser, express